#1: renamed vivoql to vivograph and get rid of secondary cmd
