+++ grapqhl_test.go
@@ -1,0 +1,3 @@
+package vivographql_test
+
+// just a placeholder now
